moved UI to Direct2D and DirectWrite, fixed String.add(), fixed event handler removing, implemented BitmapImage resource, connected UI to 3D rendering

+++ Engine/src/script/ScriptEngine.cpp
@@ -35,7 +35,7 @@ namespace Ghurund {
         return Status::OK;
     }
 
-    void ScriptEngine::update(float dt) {
+    void ScriptEngine::update(double dt) {
         for (ScriptComponent* c : Components) {
 			if (!c->Enabled)
 				continue;
